feat: creates event deeplink
