+++ src/app/[locale]/events/[state]/[eventSlug]/page.tsx
@@ -1,42 +1,47 @@
 import { Metadata } from 'next'
 import { permanentRedirect } from 'next/navigation'
 
-import { StateEventsDialog } from '@/components/app/pageEvents/components/state-events-dialog'
+import { EventDialogContent } from '@/components/app/pageEvents/components/event-dialog-content'
 import { EventsPageDialogDeeplinkLayout } from '@/components/app/pageEvents/eventsPageDialogDeeplinkLayout'
 import { PageProps } from '@/types'
+import { getEvent } from '@/utils/server/builderIO/swc-event'
 import { generateMetadataDetails } from '@/utils/server/metadataUtils'
 import { getIntlUrls } from '@/utils/shared/urls'
 import { US_MAIN_STATE_CODE_TO_DISPLAY_NAME_MAP } from '@/utils/shared/usStateUtils'
 
-type Props = PageProps<{ state: keyof typeof US_MAIN_STATE_CODE_TO_DISPLAY_NAME_MAP }>
+type Props = PageProps<{
+  state: keyof typeof US_MAIN_STATE_CODE_TO_DISPLAY_NAME_MAP
+  eventSlug: string
+}>
 
+const title = 'Events'
 const description =
   'Stand With Crypto hosts events nationwide to organize, activate, and energize the nationwide Crypto community. Check this page for information about upcoming events, including times, dates, and locations, and RSVP to events in your area'
 
-export async function generateMetadata({ params }: Props): Promise<Metadata> {
-  const { state } = params
-
-  return generateMetadataDetails({
-    title: `Events in ${US_MAIN_STATE_CODE_TO_DISPLAY_NAME_MAP[state.toUpperCase() as keyof typeof US_MAIN_STATE_CODE_TO_DISPLAY_NAME_MAP]}`,
+export const metadata: Metadata = {
+  ...generateMetadataDetails({
+    title,
     description,
-  })
+  }),
 }
 
 export default async function EventsPageRoot({ params }: Props) {
-  const { state, locale } = params
+  const { state, eventSlug, locale } = params
   const intlUrls = getIntlUrls(locale)
+
+  const event = await getEvent(eventSlug, state)
 
   const isStateValid = Object.keys(US_MAIN_STATE_CODE_TO_DISPLAY_NAME_MAP).includes(
     state.toUpperCase(),
   )
 
-  if (!isStateValid) {
+  if (!isStateValid || !event) {
     permanentRedirect(intlUrls.events())
   }
 
   return (
     <EventsPageDialogDeeplinkLayout pageParams={params}>
-      <StateEventsDialog state={state} />
+      <EventDialogContent event={event.data} />
     </EventsPageDialogDeeplinkLayout>
   )
 }
