+++ src/utils/shared/getSWCEvents.ts
@@ -5,26 +5,26 @@ export enum EVENT_TYPE_OPTIONS {
   'partner' = 'partner',
 }
 
-export const zodEventsSchemaValidation = array(
-  object({
-    data: object({
-      promotedPositioning: number().optional(),
-      image: string().url(),
-      rsvpUrl: string().url(),
-      datetime: string(),
-      formattedAddress: string(),
-      countryCode: string().length(2),
-      isOccuring: boolean(),
-      name: string(),
-      state: string().length(2),
-      type: nativeEnum(EVENT_TYPE_OPTIONS),
-      isFeatured: boolean(),
-      slug: string(),
-      city: string(),
-      description: string(),
-    }),
-    published: string().regex(/^published$/),
+export const zodEventSchemaValidation = object({
+  data: object({
+    promotedPositioning: number().optional(),
+    image: string().url(),
+    rsvpUrl: string().url(),
+    datetime: string(),
+    formattedAddress: string(),
+    countryCode: string().length(2),
+    isOccuring: boolean(),
+    name: string(),
+    state: string().length(2),
+    type: nativeEnum(EVENT_TYPE_OPTIONS),
+    isFeatured: boolean(),
+    slug: string(),
+    city: string(),
+    description: string(),
   }),
-)
+  published: string().regex(/^published$/),
+})
+
+export const zodEventsSchemaValidation = array(zodEventSchemaValidation)
 
 export type SWCEvents = Zod.infer<typeof zodEventsSchemaValidation>
